# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AgencyOS API
  description: Backend for the AI Outreach Agency Operating System.
  version: 0.1.0
paths:
...
  /api/v1/workflows:
    get:
      tags:
      - workflows
      summary: List workflows with filters
      operationId: list_workflows_api_v1_workflows_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/WorkflowStatus'
          - type: 'null'
          title: Status
      - name: sort
        in: query
        required: false
        schema:
          type: string
          pattern: ^(created_at|name|status|version)$
          default: created_at
          title: Sort
      - name: order
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          default: desc
          title: Order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workflows
      summary: Create a workflow
      operationId: create_workflow_api_v1_workflows_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflows/{workflow_id}:
    get:
      tags:
      - workflows
      summary: Get a workflow
      operationId: get_workflow_api_v1_workflows__workflow_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - workflows
      summary: Update a workflow
      operationId: update_workflow_api_v1_workflows__workflow_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - workflows
      summary: Delete a workflow
      operationId: delete_workflow_api_v1_workflows__workflow_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/workflows/{workflow_id}/executions:
    get:
      tags:
      - workflows
      summary: List executions for a workflow
      operationId: list_workflow_executions_api_v1_workflows__workflow_id__executions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflow Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: sort
        in: query
        required: false
        schema:
          type: string
          pattern: ^(created_at|started_at|finished_at|status)$
          default: created_at
          title: Sort
      - name: order
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          default: desc
          title: Order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/workflow-executions:
    post:
      tags:
      - workflow-executions
      summary: Queue a workflow execution
      operationId: queue_execution_api_v1_workflow_executions_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowExecutionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionQueue'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - workflow-executions
      summary: List workflow executions with filters
      operationId: list_executions_api_v1_workflow_executions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: workflow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Workflow Id
      - name: trigger_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trigger Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExecutionStatus'
          - type: 'null'
          title: Status
      - name: sort
        in: query
        required: false
        schema:
          type: string
          pattern: ^(created_at|started_at|finished_at|status)$
          default: created_at
          title: Sort
      - name: order
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          default: desc
          title: Order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}:
    get:
      tags:
      - workflow-executions
      summary: Get a workflow execution
      operationId: get_execution_api_v1_workflow_executions__execution_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/events:
    get:
      tags:
      - workflow-executions
      summary: Get the technical timeline for a workflow execution
      operationId: list_execution_events_api_v1_workflow_executions__execution_id__events_get
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEventListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/start:
    post:
      tags:
      - workflow-executions
      summary: Start a queued execution
      operationId: start_execution_api_v1_workflow_executions__execution_id__start_post
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/retry:
    post:
      tags:
      - workflow-executions
      summary: Retry a failed or cancelled execution
      operationId: retry_execution_api_v1_workflow_executions__execution_id__retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/complete:
    post:
      tags:
      - workflow-executions
      summary: Mark a running execution succeeded with an output payload
      operationId: complete_execution_api_v1_workflow_executions__execution_id__complete_post
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/fail:
    post:
      tags:
      - workflow-executions
      summary: Mark a running execution failed (optionally schedule a retry)
      operationId: fail_execution_api_v1_workflow_executions__execution_id__fail_post
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      - name: schedule_retry
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Schedule Retry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workflow-executions/{execution_id}/cancel:
    post:
      tags:
      - workflow-executions
      summary: Cancel a queued/running execution
      operationId: cancel_execution_api_v1_workflow_executions__execution_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecutionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ExecutionEventListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ExecutionEventRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: ExecutionEventListResponse
    ExecutionEventRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        attempt:
          type: integer
          title: Attempt
        event_type:
          $ref: '#/components/schemas/ExecutionEventType'
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - execution_id
      - attempt
      - event_type
      - metadata
      - occurred_at
      - created_at
      title: ExecutionEventRead
    ExecutionEventType:
      type: string
      enum:
      - queued
      - started
      - adapter_dispatched
      - adapter_returned
      - step_started
      - step_completed
      - step_failed
      - retrying
      - succeeded
      - failed
      - cancelled
      - timed_out
      - timeout_guard
      title: ExecutionEventType
      description: 'Closed set of append-only technical timeline labels.


        Mirrors ``public.execution_event_type`` (database/migrations/enums/).

        ``queued``/``started``/``retrying``/``succeeded``/``failed``/

        ``cancelled``/``timed_out`` are the state-machine transitions; the rest

        are worker/adapter/step instrumentation points.'
...
    ExecutionStatus:
      type: string
      enum:
      - queued
      - running
      - succeeded
      - failed
      - retrying
      - cancelled
      - timed_out
      title: ExecutionStatus
      description: Lifecycle of a workflow execution.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkflowCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        definition:
          additionalProperties: true
          type: object
          title: Definition
        execution_mode:
          type: string
          pattern: ^(n8n|builtin)$
          title: Execution Mode
          default: n8n
        config:
          additionalProperties: true
          type: object
          title: Config
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
      type: object
      required:
      - name
      title: WorkflowCreate
...
    WorkflowExecutionCreate:
      properties:
        input:
          additionalProperties: true
          type: object
          title: Input
        max_attempts:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Max Attempts
          default: 3
        retry_delay_seconds:
          type: integer
          minimum: 0.0
          title: Retry Delay Seconds
          default: 60
        retry_backoff:
          type: string
          pattern: ^(constant|exponential)$
          title: Retry Backoff
          default: exponential
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
        workflow_id:
          type: string
          format: uuid
          title: Workflow Id
        trigger_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trigger Id
        trace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trace Id
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Idempotency Key
      type: object
      required:
      - workflow_id
      title: WorkflowExecutionCreate
    WorkflowExecutionListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/WorkflowExecutionRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: WorkflowExecutionListResponse
    WorkflowExecutionQueue:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        status:
          $ref: '#/components/schemas/ExecutionStatus'
      type: object
      required:
      - execution_id
      - status
      title: WorkflowExecutionQueue
      description: Response envelope for the queue endpoint (execution_id + initial status).
    WorkflowExecutionRead:
      properties:
        input:
          additionalProperties: true
          type: object
          title: Input
        max_attempts:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Max Attempts
          default: 3
        retry_delay_seconds:
          type: integer
          minimum: 0.0
          title: Retry Delay Seconds
          default: 60
        retry_backoff:
          type: string
          pattern: ^(constant|exponential)$
          title: Retry Backoff
          default: exponential
        id:
          type: string
          format: uuid
          title: Id
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        workflow_id:
          type: string
          format: uuid
          title: Workflow Id
        trigger_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trigger Id
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        output:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        attempts:
          type: integer
          title: Attempts
        next_retry_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Next Retry At
        requested_by_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Requested By User Id
        trace_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trace Id
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        cancel_requested_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cancel Requested At
        cancelled_by_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Cancelled By User Id
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - organization_id
      - workflow_id
      - trigger_id
      - status
      - attempts
      - created_at
      - updated_at
      title: WorkflowExecutionRead
    WorkflowListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/WorkflowRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: WorkflowListResponse
    WorkflowRead:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        definition:
          additionalProperties: true
          type: object
          title: Definition
        execution_mode:
          type: string
          pattern: ^(n8n|builtin)$
          title: Execution Mode
          default: n8n
        config:
          additionalProperties: true
          type: object
          title: Config
        id:
          type: string
          format: uuid
          title: Id
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        version:
          type: integer
          title: Version
        created_by_user_id:
          type: string
          format: uuid
          title: Created By User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - name
      - id
      - organization_id
      - status
      - version
      - created_by_user_id
      - created_at
      - updated_at
      title: WorkflowRead
    WorkflowStatus:
      type: string
      enum:
      - draft
      - active
      - paused
      - archived
      title: WorkflowStatus
      description: Lifecycle of a workflow definition.
...
    WorkflowUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        definition:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Definition
        status:
          anyOf:
          - $ref: '#/components/schemas/WorkflowStatus'
          - type: 'null'
        execution_mode:
          anyOf:
          - type: string
            pattern: ^(n8n|builtin)$
          - type: 'null'
          title: Execution Mode
        config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config
      type: object
      title: WorkflowUpdate
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer